♻️ refactor : ScrollAnimations 및 AnimatedBackground 컴포넌트 애니메이션 최적화 및 MainScene 추가 #9
